Fix JavaScript errors and DeepSeek API issues
- Fix TypeError: Add null check for message.content before accessing length
- Fix ReferenceError: Move newAbortController declaration to proper scope
- Fix DeepSeek API tool_choice errors for reasoner models
- Fix assistant message format validation
- Add modular chat components and hooks
- Optimize stream chunks merging with validation
